clippy の警告を修正する
- manual_checked_ops: `remaining / bytes_per_sample` を `checked_div` に置き換える
- unnecessary_cast: u64 から u64 への不要なキャストを削除する

+++ src/mux_fmp4_segment.rs
@@ -329,7 +329,7 @@ impl Fmp4SegmentMuxer {
             )
         } else {
             // 拡張サイズが必要: ヘッダーが 16 バイトになるため合計値を再計算する
-            let extended_box_size = 16u64 + mdat_payload_size as u64;
+            let extended_box_size = 16u64 + mdat_payload_size;
             (BoxSize::U64(extended_box_size), 16)
         };
         let mdat_header = BoxHeader::new(MdatBox::TYPE, mdat_box_size);
